handle cases for issue CASSANDRA-8733
Introduces a `release_version` method to get the version of cassandra we're
connected with. This allows us to check if the version contains the fix or not.

*IMPORTANT*
This also changes the behaviour of `unshift`. The previous implementation was
compensating for the issue by reversing the list to prepend. There will be
an impact if upgrading the library w/o upgrading Cassandra to a version that
contains the fix. The reason for that is the bug was only partly compensated
by the code as shown by `data_set_spec` and `list_spec` tests.